src/docker/observation-server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Project for sandbox services observation server"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "numpy",
    "pillow",
    "pyautogui",
    "pynput",
    "requests",
    "opencv-python",
]
